introduce a reusable scanBrowserRuntimeFeatures
