--- src/internal/exploring/importJson.js
@@ -1,7 +1,0 @@
-import { fetchUrl } from "../browser-utils/fetch-browser.js"
-
-export const importJson = async (url, options = {}) => {
-  const response = await fetchUrl(url, options)
-  const object = await response.json()
-  return object
-}
